+++ core/src/main/java/cucumber/runtime/io/ResourceLoaderClassFinder.java
@@ -1,6 +1,7 @@
 package cucumber.runtime.io;
 
 import cucumber.runtime.ClassFinder;
+import io.cucumber.core.model.Classpath;
 
 import java.net.URI;
 import java.util.Collection;
@@ -8,7 +9,6 @@ import java.util.HashSet;
 
 public class ResourceLoaderClassFinder implements ClassFinder {
     private static final String CLASS_SUFFIX = ".class";
-    private static final String CLASSPATH_PREFIX = "classpath:";
     private static final char DOT = '.';
     private static final char PACKAGE_PATH_SEPARATOR = '/';
     private final ResourceLoader resourceLoader;
@@ -20,10 +20,10 @@ public class ResourceLoaderClassFinder implements ClassFinder {
     }
 
     @Override
-    public <T> Collection<Class<? extends T>> getDescendants(Class<T> parentType, String packageName) {
+    public <T> Collection<Class<? extends T>> getDescendants(Class<T> parentType, URI packageName) {
         Collection<Class<? extends T>> result = new HashSet<>();
-        for (Resource classResource : resourceLoader.resources(packagePath(packageName), CLASS_SUFFIX)) {
-            String className = getClassName(classResource.getPath().getSchemeSpecificPart());
+        for (Resource classResource : resourceLoader.resources(packageName, CLASS_SUFFIX)) {
+            String className = getClassName(classResource.getPath());
             try {
                 Class<?> clazz = loadClass(className);
                 if (clazz != null && !parentType.equals(clazz) && parentType.isAssignableFrom(clazz)) {
@@ -35,15 +35,9 @@ public class ResourceLoaderClassFinder implements ClassFinder {
         return result;
     }
 
-    private URI packagePath(String packageName) {
-        return URI.create(CLASSPATH_PREFIX + packageName.replace(DOT, PACKAGE_PATH_SEPARATOR));
-    }
-
-    private String getClassName(String path) {
-        if (path.charAt(0) == PACKAGE_PATH_SEPARATOR) {
-            path = path.substring(1);
-        }
-        return path.substring(0, path.length() - CLASS_SUFFIX.length()).replace(PACKAGE_PATH_SEPARATOR, DOT);
+    private static String getClassName(URI uri) {
+        String resourceName = Classpath.resourceName(uri);
+        return resourceName.substring(0, resourceName.length() - CLASS_SUFFIX.length()).replace(PACKAGE_PATH_SEPARATOR, DOT);
     }
 
     public <T> Class<? extends T> loadClass(String className) throws ClassNotFoundException {