+++ core/src/main/java/cucumber/runtime/model/FeatureLoader.java
@@ -2,8 +2,8 @@ package cucumber.runtime.model;
 
 import cucumber.runtime.io.Resource;
 import cucumber.runtime.io.ResourceLoader;
+import io.cucumber.core.model.FeatureIdentifier;
 
-import java.io.PrintStream;
 import java.net.URI;
 import java.util.Iterator;
 import java.util.List;
@@ -15,18 +15,6 @@ public final class FeatureLoader {
 
     public FeatureLoader(ResourceLoader resourceLoader) {
         this.resourceLoader = resourceLoader;
-    }
-
-    public List<CucumberFeature> load(List<URI> featurePaths, PrintStream out) {
-        final List<CucumberFeature> cucumberFeatures = load(featurePaths);
-        if (cucumberFeatures.isEmpty()) {
-            if (featurePaths.isEmpty()) {
-                out.println("Got no path to feature directory or feature file");
-            } else {
-                out.println(String.format("No features found at %s", featurePaths));
-            }
-        }
-        return cucumberFeatures;
     }
 
     public List<CucumberFeature> load(List<URI> featurePaths) {
@@ -41,7 +29,7 @@ public final class FeatureLoader {
         Iterable<Resource> resources = resourceLoader.resources(featurePath, FEATURE_SUFFIX);
 
         Iterator<Resource> iterator = resources.iterator();
-        if (featurePath.getSchemeSpecificPart().endsWith(FEATURE_SUFFIX) && !iterator.hasNext()) {
+        if (FeatureIdentifier.isFeature(featurePath) && !iterator.hasNext()) {
             throw new IllegalArgumentException("Feature not found: " + featurePath);
         }
         while (iterator.hasNext()) {