+++ core/src/test/java/cucumber/runtime/BackendModuleBackendSupplierTest.java
@@ -7,8 +7,10 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
+import java.net.URI;
 import java.util.Collections;
 
+import static java.util.Collections.singletonList;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.notNullValue;
 import static org.junit.Assert.assertThat;
@@ -35,7 +37,7 @@ public class BackendModuleBackendSupplierTest {
         RuntimeOptions runtimeOptions = new RuntimeOptions(Collections.<String>emptyList());
         ResourceLoader resourceLoader = new MultiLoader(classLoader);
         ClassFinder classFinder = new ResourceLoaderClassFinder(resourceLoader, classLoader);
-        BackendSupplier backendSupplier = new BackendModuleBackendSupplier(resourceLoader, classFinder, runtimeOptions, Collections.singletonList("no.backend.here"));
+        BackendSupplier backendSupplier = new BackendModuleBackendSupplier(resourceLoader, classFinder, runtimeOptions, singletonList(URI.create("classpath:no/backend/here")));
 
         expectedException.expect(CucumberException.class);
         assertThat(backendSupplier.get().iterator().next(), is(notNullValue()));