+++ core/src/test/java/cucumber/runtime/RuntimeOptionsFactoryTest.java
@@ -14,8 +14,6 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.regex.Pattern;
 
-import static cucumber.runtime.RuntimeOptionsFactory.packageName;
-import static cucumber.runtime.RuntimeOptionsFactory.packagePath;
 import static java.util.Arrays.asList;
 import static org.hamcrest.collection.IsCollectionWithSize.hasSize;
 import static org.hamcrest.collection.IsIterableContainingInOrder.contains;
@@ -101,16 +99,6 @@ public class RuntimeOptionsFactoryTest {
 
     private String getRegexpPattern(Object pattern) {
         return ((Pattern) pattern).pattern();
-    }
-
-    @Test
-    public void finds_path_for_class_in_package() {
-        assertEquals("java/lang", packagePath(String.class));
-    }
-
-    @Test
-    public void finds_path_for_class_in_toplevel_package() {
-        assertEquals("", packageName("TopLevelClass"));
     }
 
     @Test